pyproject.toml
[tool.black]
line-length = 120

[tool.poetry]
name = "sporthalle"
version = "0.1.0"
description = "pull events from sporthalle.hamburg/terminkalender and push them to a nextcloud calendar"
authors = ["Eulentier161 <git@eule.wtf>"]
readme = "README.md"
license = "GPL-3.0-or-later"

[tool.poetry.dependencies]
python = "^3.12"
httpx = "^0.27.2"
beautifulsoup4 = "^4.12.3"
python-dotenv = "^1.0.1"
caldav = "^1.4.0"

[tool.poetry.scripts]
crawl = "sporthalle.crawl:main"

[tool.poetry.group.dev.dependencies]
black = "^24.10.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
